refactor: rename passenger to customer to match actual role
- Rename app/(passenger) to app/(customer)
- Rename scenes/passenger to scenes/customer
- Update all imports and function names from Passenger* to Customer*
- Update routing to use /(customer) instead of /(passenger)
- Remove 'PASSENGER' fallback from routing logic

This aligns the folder structure and naming with the actual Customer role from the backend.
