+++ app/(customer)/_layout.tsx
@@ -1,0 +1,50 @@
+import { Tabs } from 'expo-router';
+import { useAppTheme } from '@/plugin/theme-provider';
+import { colors } from '@/theme';
+import { Ionicons } from '@expo/vector-icons';
+
+export default function CustomerTabsLayout() {
+  const { isDark } = useAppTheme();
+
+  return (
+    <Tabs
+      screenOptions={{
+        headerShown: false,
+        tabBarActiveTintColor: colors.darkPurple,
+        tabBarInactiveTintColor: isDark ? colors.gray : colors.darkGray,
+        tabBarStyle: {
+          backgroundColor: isDark ? colors.blackGray : colors.white,
+          borderTopColor: isDark ? colors.darkGray : colors.lightGray,
+        },
+      }}>
+      <Tabs.Screen
+        name="home"
+        options={{
+          title: 'Home',
+          tabBarIcon: ({ color, size }) => <Ionicons name="home" size={size} color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="trips"
+        options={{
+          title: 'Trips',
+          tabBarIcon: ({ color, size }) => <Ionicons name="car" size={size} color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="wallet"
+        options={{
+          title: 'Wallet',
+          tabBarIcon: ({ color, size }) => <Ionicons name="wallet" size={size} color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="profile"
+        options={{
+          title: 'Profile',
+          tabBarIcon: ({ color, size }) => <Ionicons name="person" size={size} color={color} />,
+        }}
+      />
+    </Tabs>
+  );
+}
