+++ app/(customer)/home/book-delivery.tsx
@@ -1,0 +1,3 @@
+import BookDelivery from '@/scenes/customer/home/BookDelivery';
+
+export default BookDelivery;
