+++ app/(customer)/home/book-ride.tsx
@@ -1,0 +1,3 @@
+import BookRide from '@/scenes/customer/home/BookRide';
+
+export default BookRide;
