+++ app/(customer)/home/index.tsx
@@ -1,0 +1,3 @@
+import CustomerHome from '@/scenes/customer/home/CustomerHome';
+
+export default CustomerHome;
