+++ app/(customer)/profile/edit-profile.tsx
@@ -1,0 +1,3 @@
+import EditProfile from '@/scenes/customer/profile/EditProfile';
+
+export default EditProfile;
