+++ app/(customer)/profile/index.tsx
@@ -1,0 +1,3 @@
+import CustomerProfile from '@/scenes/customer/profile/CustomerProfile';
+
+export default CustomerProfile;
