+++ app/(customer)/profile/settings.tsx
@@ -1,0 +1,3 @@
+import CustomerSettings from '@/scenes/customer/profile/CustomerSettings';
+
+export default CustomerSettings;
