+++ app/(customer)/profile/support.tsx
@@ -1,0 +1,3 @@
+import Support from '@/scenes/customer/profile/Support';
+
+export default Support;
