+++ app/(customer)/trips/index.tsx
@@ -1,0 +1,3 @@
+import CustomerTrips from '@/scenes/customer/trips/CustomerTrips';
+
+export default CustomerTrips;
