+++ app/(customer)/trips/tracking.tsx
@@ -1,0 +1,3 @@
+import TripTracking from '@/scenes/customer/trips/TripTracking';
+
+export default TripTracking;
