+++ app/(customer)/trips/trip-details.tsx
@@ -1,0 +1,3 @@
+import TripDetails from '@/scenes/customer/trips/TripDetails';
+
+export default TripDetails;
