+++ app/(customer)/wallet/add-payment-method.tsx
@@ -1,0 +1,3 @@
+import AddPaymentMethod from '@/scenes/customer/wallet/AddPaymentMethod';
+
+export default AddPaymentMethod;
