+++ app/(customer)/wallet/index.tsx
@@ -1,0 +1,3 @@
+import CustomerWallet from '@/scenes/customer/wallet/CustomerWallet';
+
+export default CustomerWallet;
