+++ app/(customer)/wallet/payment-methods.tsx
@@ -1,0 +1,3 @@
+import PaymentMethods from '@/scenes/customer/wallet/PaymentMethods';
+
+export default PaymentMethods;
