+++ app/(customer)/wallet/transaction-history.tsx
@@ -1,0 +1,3 @@
+import TransactionHistory from '@/scenes/customer/wallet/TransactionHistory';
+
+export default TransactionHistory;
