--- app/(passenger)/home/book-delivery.tsx
@@ -1,3 +1,0 @@
-import BookDelivery from '@/scenes/passenger/home/BookDelivery';
-
-export default BookDelivery;
