--- app/(passenger)/home/book-ride.tsx
@@ -1,3 +1,0 @@
-import BookRide from '@/scenes/passenger/home/BookRide';
-
-export default BookRide;
