--- app/(passenger)/home/index.tsx
@@ -1,3 +1,0 @@
-import PassengerHome from '@/scenes/passenger/home/PassengerHome';
-
-export default PassengerHome;
