--- app/(passenger)/profile/_layout.tsx
@@ -1,22 +1,0 @@
-import { Stack } from 'expo-router';
-import { useAppTheme } from '@/plugin/theme-provider';
-import { colors } from '@/theme';
-
-export default function PassengerProfileStackLayout() {
-  const { isDark } = useAppTheme();
-
-  return (
-    <Stack
-      screenOptions={{
-        headerTintColor: colors.white,
-        headerStyle: { backgroundColor: isDark ? colors.blackGray : colors.darkPurple },
-        headerTitleStyle: { fontSize: 18 },
-        headerTitleAlign: 'center',
-      }}>
-      <Stack.Screen name="index" options={{ title: 'Profile' }} />
-      <Stack.Screen name="edit-profile" options={{ title: 'Edit Profile' }} />
-      <Stack.Screen name="settings" options={{ title: 'Settings' }} />
-      <Stack.Screen name="support" options={{ title: 'Support' }} />
-    </Stack>
-  );
-}
