--- app/(passenger)/profile/edit-profile.tsx
@@ -1,3 +1,0 @@
-import EditProfile from '@/scenes/passenger/profile/EditProfile';
-
-export default EditProfile;
