--- app/(passenger)/profile/index.tsx
@@ -1,3 +1,0 @@
-import PassengerProfile from '@/scenes/passenger/profile/PassengerProfile';
-
-export default PassengerProfile;
