--- app/(passenger)/profile/settings.tsx
@@ -1,3 +1,0 @@
-import PassengerSettings from '@/scenes/passenger/profile/PassengerSettings';
-
-export default PassengerSettings;
