--- app/(passenger)/profile/support.tsx
@@ -1,3 +1,0 @@
-import Support from '@/scenes/passenger/profile/Support';
-
-export default Support;
