--- app/(passenger)/trips/index.tsx
@@ -1,3 +1,0 @@
-import PassengerTrips from '@/scenes/passenger/trips/PassengerTrips';
-
-export default PassengerTrips;
