--- app/(passenger)/trips/tracking.tsx
@@ -1,3 +1,0 @@
-import TripTracking from '@/scenes/passenger/trips/TripTracking';
-
-export default TripTracking;
