--- app/(passenger)/trips/trip-details.tsx
@@ -1,3 +1,0 @@
-import TripDetails from '@/scenes/passenger/trips/TripDetails';
-
-export default TripDetails;
