--- app/(passenger)/wallet/_layout.tsx
@@ -1,22 +1,0 @@
-import { Stack } from 'expo-router';
-import { useAppTheme } from '@/plugin/theme-provider';
-import { colors } from '@/theme';
-
-export default function PassengerWalletStackLayout() {
-  const { isDark } = useAppTheme();
-
-  return (
-    <Stack
-      screenOptions={{
-        headerTintColor: colors.white,
-        headerStyle: { backgroundColor: isDark ? colors.blackGray : colors.darkPurple },
-        headerTitleStyle: { fontSize: 18 },
-        headerTitleAlign: 'center',
-      }}>
-      <Stack.Screen name="index" options={{ title: 'Wallet' }} />
-      <Stack.Screen name="payment-methods" options={{ title: 'Payment Methods' }} />
-      <Stack.Screen name="transaction-history" options={{ title: 'Transaction History' }} />
-      <Stack.Screen name="add-payment-method" options={{ title: 'Add Payment Method' }} />
-    </Stack>
-  );
-}
