--- app/(passenger)/wallet/add-payment-method.tsx
@@ -1,3 +1,0 @@
-import AddPaymentMethod from '@/scenes/passenger/wallet/AddPaymentMethod';
-
-export default AddPaymentMethod;
