--- app/(passenger)/wallet/index.tsx
@@ -1,3 +1,0 @@
-import PassengerWallet from '@/scenes/passenger/wallet/PassengerWallet';
-
-export default PassengerWallet;
