--- app/(passenger)/wallet/payment-methods.tsx
@@ -1,3 +1,0 @@
-import PaymentMethods from '@/scenes/passenger/wallet/PaymentMethods';
-
-export default PaymentMethods;
