--- app/(passenger)/wallet/transaction-history.tsx
@@ -1,3 +1,0 @@
-import TransactionHistory from '@/scenes/passenger/wallet/TransactionHistory';
-
-export default TransactionHistory;
