+++ scenes/customer/profile/CustomerSettings.tsx
@@ -1,0 +1,32 @@
+import { StyleSheet, Text, View } from 'react-native';
+import { colors } from '@/theme';
+import { useAppTheme } from '@/plugin/theme-provider';
+
+const styles = StyleSheet.create({
+  root: {
+    flex: 1,
+    padding: 20,
+    backgroundColor: colors.lightGrayPurple,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 20,
+  },
+  description: {
+    fontSize: 16,
+  },
+});
+
+export default function CustomerSettings() {
+  const { isDark } = useAppTheme();
+
+  return (
+    <View style={[styles.root, isDark && { backgroundColor: colors.blackGray }]}>
+      <Text style={[styles.title, isDark && { color: colors.white }]}>Settings</Text>
+      <Text style={[styles.description, isDark && { color: colors.gray }]}>
+        Manage app preferences, notifications, privacy, and security settings.
+      </Text>
+    </View>
+  );
+}
