+++ scenes/customer/trips/TripDetails.tsx
